Refactored state_transition
added missing method comments
